BIMind/requirements.txt
fastapi
uvicorn
ifcopenshell
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
